chore: run desk rename codemod
