+++ dev/design-studio/sanity.config.ts
@@ -1,5 +1,5 @@
 import {defineConfig} from 'sanity'
-import {deskTool} from 'sanity/desk'
+import {structureTool} from 'sanity/structure'
 import {templates} from './templates'
 import {themePreviewTool} from './plugins/theme-preview-tool'
 import {schemaTypes} from './schemaTypes'
@@ -7,7 +7,7 @@ import {structure} from './structure'
 
 export default defineConfig({
   plugins: [
-    deskTool({structure}),
+    structureTool({structure}),
     themePreviewTool(),
     // @todo
     // visionTool(),
