+++ dev/design-studio/structure.ts
@@ -1,4 +1,4 @@
-import {StructureResolver, DefaultDocumentNodeResolver} from 'sanity/desk'
+import {StructureResolver, DefaultDocumentNodeResolver} from 'sanity/structure'
 import {CogIcon} from '@sanity/icons'
 import {CustomPane} from './components/CustomPane'
 import {IFrameView} from './components/IframeView'
