+++ dev/test-studio/components/panes/debug/DebugPane.tsx
@@ -1,4 +1,4 @@
-import {UserComponent, usePaneRouter} from 'sanity/desk'
+import {UserComponent, usePaneRouter} from 'sanity/structure'
 import {ChevronDownIcon, ChevronRightIcon, ControlsIcon, LinkIcon} from '@sanity/icons'
 import {Box, Card, Code, Flex, Stack, Text} from '@sanity/ui'
 import React, {useMemo} from 'react'
