+++ dev/test-studio/structure/resolveStructureDocumentNode.ts
@@ -1,4 +1,4 @@
-import {DefaultDocumentNodeResolver} from 'sanity/desk'
+import {DefaultDocumentNodeResolver} from 'sanity/structure'
 import {JSONPreviewDocumentView} from '../components/documentViews/jsonPreview'
 
 export const defaultDocumentNode: DefaultDocumentNodeResolver = (S, {schemaType}) => {
