+++ packages/@sanity/cli/templates/shopify-online-storefront/structure/productStructure.ts
@@ -1,5 +1,5 @@
-import {ListItemBuilder} from 'sanity/desk'
 import {InfoOutlineIcon} from '@sanity/icons'
+import {ListItemBuilder} from 'sanity/structure'
 import defineStructure from '../utils/defineStructure'
 
 export default defineStructure<ListItemBuilder>((S) =>
