+++ packages/@sanity/cli/templates/shopify-online-storefront/utils/defineStructure.ts
@@ -1,5 +1,5 @@
 import {ConfigContext} from 'sanity'
-import {StructureBuilder} from 'sanity/desk'
+import {StructureBuilder} from 'sanity/structure'
 
 /**
  * Helper for creating and typing composable structure parts.
