+++ packages/@sanity/cli/templates/shopify/structure/pageStructure.ts
@@ -1,6 +1,6 @@
-import {ListItemBuilder} from 'sanity/desk'
-import defineStructure from '../utils/defineStructure'
 import {DocumentsIcon} from '@sanity/icons'
+import {ListItemBuilder} from 'sanity/structure';
+import defineStructure from '../utils/defineStructure'
 
 export default defineStructure<ListItemBuilder>((S) =>
   S.listItem()
