+++ packages/@sanity/cli/test/__fixtures__/v3/sanity.config.ts
@@ -1,5 +1,5 @@
 import {defineConfig} from 'sanity'
-import {deskTool} from 'sanity/desk'
+import {structureTool} from 'sanity/structure'
 import {MyLogo} from './components/MyLogo'
 import {schema} from './schema'
 
@@ -7,7 +7,7 @@ export default defineConfig({
   projectId: process.env.SANITY_STUDIO_PROJECT_ID || 'missingenv',
   apiHost: 'https://api.sanity.work',
   dataset: process.env.SANITY_STUDIO_DATASET || 'missingenv',
-  plugins: [deskTool()],
+  plugins: [structureTool()],
   logo: MyLogo,
   schema,
 })
